refactor: Rename `…s/aws/package/compile/events/alb/lib/targetGroups.js`
