+++ lib/plugins/aws/package/compile/events/alb/index.js
@@ -3,7 +3,7 @@
 const BbPromise = require('bluebird');
 
 const validate = require('./lib/validate');
-const compileTargetGroups = require('./lib/targetGroups');
+const compileTargetGroups = require('./lib/target-groups');
 const compileListenerRules = require('./lib/listenerRules');
 const compilePermissions = require('./lib/permissions');
